pyproject.toml
[tool.poetry]
name = "third-eye"
version = "0.1.0"
description = ""
authors = ["ruthus18 <ruthus18@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.16.1"
APScheduler = "^3.7.0"
pytz = "^2020.4"
pydantic = {version = "^1.8", extras = ["dotenv"]}
tortoise-orm = "^0.16.21"
aerich = "^0.5.0"
aiocron = "^1.4"
asyncpg = "^0.22.0"

[tool.poetry.dev-dependencies]
ipython = "^7.20.0"
isort = "^5.7.0"
flake8 = "^3.8.4"
mypy = "^0.812"
pytest = "^6.2.2"
pytest-asyncio = "^0.14.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
